requirements.txt
python-telegram-bot==21.4
gspread==6.1.3
google-auth==2.29.0
python-dotenv==1.0.1
